==> notebooks/02_eda_amazon_books.ipynb <==
### 02 – EDA on Amazon Books Dataset

Goal:
- Load an Amazon Books subset.
- Confirm presence of user_id, item_id, rating, timestamp.
- Map to unified schema.


import pandas as pd
from pathlib import Path


# download dataset from https://mcauleylab.ucsd.edu/public_datasets/data/amazon_2023/raw/review_categories/Books.jsonl.gz
path = "..\\data\\raw\\amazon_books\\Books.jsonl.gz"
sample_df = pd.read_json(
    path,
    lines=True,
    compression="gzip",
    nrows=100_000  # just to inspect columns
)

print(sample_df.shape)
print(sample_df.columns)
sample_df.head()


#df_full = pd.read_json(path, lines=True, compression="gzip")

user_col = "user_id"
item_col = "parent_asin"
rating_col = "rating"
time_col = "timestamp"

df = sample_df[[user_col, item_col, rating_col, time_col]].copy()

# Keep only positive interactions (e.g., rating >= 3)
df = df[df[rating_col] >= 3].copy()

# Parse timestamp
df["timestamp"] = pd.to_datetime(df[time_col], errors="coerce")
df = df.dropna(subset=["timestamp"])
df = df.sort_values([user_col, "timestamp"])

df_uni = pd.DataFrame({
    "dataset": "amazon_books_2023",
    "user_id": df[user_col].astype(str),
    "session_id": df[user_col].astype(str),   # placeholder; sessions later
    "item_id": df[item_col].astype(str),
    "timestamp": df["timestamp"],
    "interaction_type": "implicit_pos"
})

df_uni.head()


path = Path("../data/raw/amazon_books/Books.jsonl.gz")
out_path = Path("../data/processed/amazon_books_2023_interactions.parquet")

# If rerunning, delete old file
if out_path.exists():
    out_path.unlink()

chunks = pd.read_json(
    path,
    lines=True,
    compression="gzip",
    chunksize=500_000  # adjust to your RAM
)

dfs = []
for i, chunk in enumerate(chunks):
    print(f"Processing chunk {i}...")
    user_col = "user_id"       # adjust if needed
    item_col = "parent_asin"
    rating_col = "rating"
    time_col = "timestamp"

    df = chunk[[user_col, item_col, rating_col, time_col]].copy()
    df = df[df[rating_col] >= 3].copy()
    df["timestamp"] = pd.to_datetime(df[time_col], errors="coerce")
    df = df.dropna(subset=["timestamp"])
    df = df.sort_values([user_col, "timestamp"])

    df_uni = pd.DataFrame({
        "dataset": "amazon_books_2023",
        "user_id": df[user_col].astype(str),
        "session_id": df[user_col].astype(str),
        "item_id": df[item_col].astype(str),
        "timestamp": df["timestamp"],
        "interaction_type": "implicit_pos"
    })

    dfs.append(df_uni)

df_all = pd.concat(dfs, ignore_index=True)
df_all.to_parquet(out_path, index=False)
print("Saved unified interactions to:", out_path)


df_all.head()

sample_df.shape

df_all.shape

df_all.columns
